Add: client_infos.hpp and server_infos.hpp in http
