+++ srcs/http/client_infos.hpp
@@ -1,0 +1,16 @@
+#ifndef CLIENT_INFOS_HPP_
+#define CLIENT_INFOS_HPP_
+
+#include <string>
+
+namespace http {
+
+struct MockDtoClientInfos {
+	int         fd;
+	std::string request_buf;
+	std::string ip;
+};
+
+} // namespace http
+
+#endif
