+++ srcs/http/tmp_http.hpp
@@ -32,9 +32,8 @@ class TmpHttp {
   public:
 	TmpHttp();
 	~TmpHttp();
-	HttpResult
-		 Run(const server::DtoClientInfos &client_info, const server::DtoServerInfos &server_info);
-	void ParseHttpRequestFormat(int client_fd, const std::string &read_buf);
+	HttpResult Run(const MockDtoClientInfos &client_info, const MockDtoServerInfos &server_info);
+	void       ParseHttpRequestFormat(int client_fd, const std::string &read_buf);
 	utils::Result<int> TmpParseHttpRequestFormat(int client_fd, const std::string &read_buf);
 	std::string        CreateHttpResponse(int client_fd);
 	// todo: 408のtimeoutのレスポンス
